requirements.txt
Flask>=3.0,<4.0
zeroc-ice>=3.7,<3.8
